Create README.md
I install Big Sur11.3 on iMac in late 2010 and some things don't work right, the sleep button is not working, and minimize the screen, otherwise, everything else is working really well. I had my iMac for a few years and since then, I install Yosemite, High Sierra, Mojave, Catalina, and now with the new Patcher Big Sur 11.3, is an old computer, but I think Apple did an excellent job with this model. Is the challenge it makes me in control of whatever new software.